+++ opentelemetry-exporter-gcp-trace/CHANGELOG.md
@@ -5,6 +5,9 @@
 - Add optional resource attributes to trace spans with regex
   ([#145](https://github.com/GoogleCloudPlatform/opentelemetry-operations-python/pull/145))
 - Upgrade `google-cloud-trace` dependency to version 1.1 or newer.
+  ([#170](https://github.com/GoogleCloudPlatform/opentelemetry-operations-python/pull/170))
+- Fix span attribute value truncation
+  ([#173](https://github.com/GoogleCloudPlatform/opentelemetry-operations-python/pull/173))
 
 ## Version 1.0.0