+++ opentelemetry-exporter-gcp-trace/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## Version 1.1.0
+
+Released 2022-01-13
+
 - Add optional resource attributes to trace spans with regex
   ([#145](https://github.com/GoogleCloudPlatform/opentelemetry-operations-python/pull/145))
 - Upgrade `google-cloud-trace` dependency to version 1.1 or newer.